Revert MongoDB commits and add details in MongoDB update script
