+++ src/main/java/io/antmedia/datastore/db/MongoStore.java
@@ -38,7 +38,6 @@ import dev.morphia.query.filters.Filters;
 import dev.morphia.query.updates.UpdateOperator;
 import dev.morphia.query.updates.UpdateOperators;
 import io.antmedia.AntMediaApplicationAdapter;
-import io.antmedia.ScriptRunUtil;
 import io.antmedia.datastore.db.types.Broadcast;
 import io.antmedia.datastore.db.types.ConferenceRoom;
 import io.antmedia.datastore.db.types.Endpoint;
@@ -78,21 +77,14 @@ public class MongoStore extends DataStore {
 	private static final String HLS_VIEWER_COUNT = "hlsViewerCount";
 	private static final String WEBRTC_VIEWER_COUNT = "webRTCViewerCount";
 	private static final String META_DATA = "metaData";
-	public boolean dbUpdated = false;
-	
-	public boolean isDbUpdated() {
-		return dbUpdated;
-	}
-
-	ScriptRunUtil scriptUtil = new ScriptRunUtil();
 
 	public MongoStore(String host, String username, String password, String dbName) {
 
 		String uri = getMongoConnectionUri(host, username, password);
 
+
 		mongoClient = MongoClients.create(uri);
-		
-		updateMongoDb(host, username, password);
+
 
 		//TODO: Refactor these stores so that we don't have separate datastore for each class
 		datastore = Morphia.createDatastore(mongoClient, dbName);
@@ -142,14 +134,6 @@ public class MongoStore extends DataStore {
 		logger.info("uri:{}",uri);
 
 		return uri;
-	}
-	
-	public void updateMongoDb(String dbHost, String dbUser, String dbPassword) {
-		synchronized(this) {
-			dbUpdated = true;
-			String command = "/bin/bash update_mongodb.sh " + dbHost + " "+ dbUser + " " + dbPassword;
-			scriptUtil.runCommand(command);
-		}
 	}
 
 	/*
